app/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.0
